# website/hugo/static/openapi.yaml
openapi: 3.1.0
info:
    title: Taskschmiede REST API
    version: v1
    description: |
        Agent-first task and project management API. All endpoints return JSON.
        Timestamps are UTC in RFC 3339 format. Entity IDs use the format `prefix_hexstring`
        (e.g., `tsk_a1b2c3...`, `edv_d4e5f6...`).
    license:
        name: Apache 2.0
        url: https://www.apache.org/licenses/LICENSE-2.0
    contact:
        name: Quest Financial Technologies
        url: https://github.com/QuestFinTech/taskschmiede
servers:
    - url: http://localhost:9000
      description: Local development
    - url: https://api.taskschmiede.com.home.arpa
      description: Staging
security:
    - bearerAuth: []
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            description: Token from POST /api/v1/auth/login
...
    schemas:
...
        LoginRequest:
            type: object
            required: [email, password]
            properties:
                email: {type: string, format: email}
                password: {type: string}
        LoginResponse:
            type: object
            properties:
                token: {type: string}
                token_id: {type: string}
                user_id: {type: string}
                name: {type: string}
                email: {type: string}
                expires_at: {type: string, format: date-time}
...
    responses:
...
        Unauthorized:
            description: Missing or invalid authentication
            content:
                application/json:
                    schema: {$ref: "#/components/schemas/Error"}
...
paths:
...
    /api/v1/auth/login:
        post:
            tags: [Auth]
            summary: Login
            security: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema: {$ref: "#/components/schemas/LoginRequest"}
            responses:
                "200":
                    description: Login successful
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data: {$ref: "#/components/schemas/LoginResponse"}
                "401": {$ref: "#/components/responses/Unauthorized"}
    /api/v1/auth/register:
        post:
            tags: [Auth]
            summary: Register new user
            security: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema: {$ref: "#/components/schemas/RegisterRequest"}
            responses:
                "201":
                    description: Registration initiated (verification email sent)
                "400": {$ref: "#/components/responses/BadRequest"}
                "409":
                    description: Email already registered
                    content:
                        application/json:
                            schema: {$ref: "#/components/schemas/Error"}
    /api/v1/auth/verify:
        post:
            tags: [Auth]
            summary: Verify email with code
            security: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema: {$ref: "#/components/schemas/VerifyRequest"}
            responses:
                "200":
                    description: Email verified, account activated
                "400": {$ref: "#/components/responses/BadRequest"}
    /api/v1/auth/forgot-password:
        post:
            tags: [Auth]
            summary: Request password reset
            security: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema: {$ref: "#/components/schemas/ForgotPasswordRequest"}
            responses:
                "200":
                    description: Reset code sent (same response whether email exists or not)
    /api/v1/auth/reset-password:
        post:
            tags: [Auth]
            summary: Reset password with code
            security: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema: {$ref: "#/components/schemas/ResetPasswordRequest"}
            responses:
                "200":
                    description: Password reset successful
                "400": {$ref: "#/components/responses/BadRequest"}
    /api/v1/auth/resend-verification:
        post:
            tags: [Auth]
            summary: Resend verification code
            security: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [email]
                            properties:
                                email: {type: string, format: email}
            responses:
                "200":
                    description: Verification code resent
    # ──────────────────────────────────────────────
    # Auth (Authenticated)
    # ──────────────────────────────────────────────
    /api/v1/auth/whoami:
        get:
            tags: [Auth]
            summary: Get current user profile with tier info
            responses:
                "200":
                    description: User profile
                "401": {$ref: "#/components/responses/Unauthorized"}
    /api/v1/auth/profile:
        patch:
            tags: [Auth]
            summary: Update profile
            requestBody:
                required: true
                content:
                    application/json:
                        schema: {$ref: "#/components/schemas/ProfileUpdate"}
            responses:
                "200":
                    description: Profile updated
                "401": {$ref: "#/components/responses/Unauthorized"}
    /api/v1/auth/change-password:
        post:
            tags: [Auth]
            summary: Change password
            requestBody:
                required: true
                content:
                    application/json:
                        schema: {$ref: "#/components/schemas/ChangePasswordRequest"}
            responses:
                "200":
                    description: Password changed
                "400": {$ref: "#/components/responses/BadRequest"}
...
    /api/v1/tasks:
        post:
            tags: [Tasks]
            summary: Create task
            requestBody:
                required: true
                content:
                    application/json:
                        schema: {$ref: "#/components/schemas/TaskCreate"}
            responses:
                "201":
                    description: Task created
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data: {$ref: "#/components/schemas/Task"}
                "400": {$ref: "#/components/responses/BadRequest"}
        get:
            tags: [Tasks]
            summary: List tasks
            parameters:
                - {$ref: "#/components/parameters/limitParam"}
                - {$ref: "#/components/parameters/offsetParam"}
                - name: status
                  in: query
                  schema: {type: string, enum: [planned, active, done, canceled]}
                - name: endeavour_id
                  in: query
                  schema: {type: string}
                - name: assignee_id
                  in: query
                  schema: {type: string}
                  description: "Use 'me' to filter by current user"
                - name: demand_id
                  in: query
                  schema: {type: string}
                - name: search
                  in: query
                  schema: {type: string}
                - name: unassigned
                  in: query
                  schema: {type: string, enum: ["true"]}
                - name: summary
                  in: query
                  schema: {type: string, enum: ["true"]}
                  description: "Return status counts instead of list"
            responses:
                "200":
                    description: Task list
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        type: array
                                        items: {$ref: "#/components/schemas/Task"}
                                    meta: {$ref: "#/components/schemas/ListMeta"}
    /api/v1/tasks/{id}:
        get:
            tags: [Tasks]
            summary: Get task
            parameters:
                - {$ref: "#/components/parameters/idPath"}
            responses:
                "200":
                    description: Task details
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data: {$ref: "#/components/schemas/Task"}
                "404": {$ref: "#/components/responses/NotFound"}
        patch:
            tags: [Tasks]
            summary: Update task
            parameters:
                - {$ref: "#/components/parameters/idPath"}
            requestBody:
                required: true
                content:
                    application/json:
                        schema: {$ref: "#/components/schemas/TaskUpdate"}
            responses:
                "200":
                    description: Task updated
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data: {$ref: "#/components/schemas/Task"}
                "400": {$ref: "#/components/responses/BadRequest"}
                "404": {$ref: "#/components/responses/NotFound"}
    /api/v1/tasks/{id}/dod-check:
        post:
            tags: [DoD]
            summary: Check task DoD compliance
            parameters:
                - {$ref: "#/components/parameters/idPath"}
            responses:
                "200":
                    description: DoD check result
    /api/v1/tasks/{id}/dod-override:
        post:
            tags: [DoD]
            summary: Override DoD for task
            parameters:
                - {$ref: "#/components/parameters/idPath"}
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [reason]
                            properties:
                                reason: {type: string}
            responses:
                "200":
                    description: DoD overridden
    # ──────────────────────────────────────────────
    # Endeavours
    # ──────────────────────────────────────────────
    /api/v1/endeavours:
        post:
            tags: [Endeavours]
            summary: Create endeavour
            requestBody:
                required: true
                content:
                    application/json:
                        schema: {$ref: "#/components/schemas/EndeavourCreate"}
            responses:
                "201":
                    description: Endeavour created
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data: {$ref: "#/components/schemas/Endeavour"}
        get:
            tags: [Endeavours]
            summary: List endeavours
            parameters:
                - {$ref: "#/components/parameters/limitParam"}
                - {$ref: "#/components/parameters/offsetParam"}
                - name: status
                  in: query
                  schema: {type: string}
                - name: organization_id
                  in: query
                  schema: {type: string}
                - name: search
                  in: query
                  schema: {type: string}
            responses:
                "200":
                    description: Endeavour list
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        type: array
                                        items: {$ref: "#/components/schemas/Endeavour"}
                                    meta: {$ref: "#/components/schemas/ListMeta"}
    /api/v1/endeavours/{id}:
        get:
            tags: [Endeavours]
            summary: Get endeavour
            parameters:
                - {$ref: "#/components/parameters/idPath"}
            responses:
                "200":
                    description: Endeavour details
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data: {$ref: "#/components/schemas/Endeavour"}
                "404": {$ref: "#/components/responses/NotFound"}
        patch:
            tags: [Endeavours]
            summary: Update endeavour
            parameters:
                - {$ref: "#/components/parameters/idPath"}
            requestBody:
                required: true
                content:
                    application/json:
                        schema: {$ref: "#/components/schemas/EndeavourUpdate"}
            responses:
                "200":
                    description: Endeavour updated
    /api/v1/endeavours/{id}/archive:
        get:
            tags: [Endeavours]
            summary: Preview archive impact
            parameters:
                - {$ref: "#/components/parameters/idPath"}
            responses:
                "200":
                    description: Archive impact summary
        post:
            tags: [Endeavours]
            summary: Archive endeavour
            parameters:
                - {$ref: "#/components/parameters/idPath"}
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                reason: {type: string}
            responses:
                "200":
                    description: Endeavour archived
    /api/v1/endeavours/{id}/export:
        get:
            tags: [Endeavours]
            summary: Export endeavour data
            parameters:
                - {$ref: "#/components/parameters/idPath"}
            responses:
                "200":
                    description: Full endeavour export (JSON)
    /api/v1/endeavours/{id}/dod-policy:
        post:
            tags: [DoD]
            summary: Assign DoD policy to endeavour
            parameters:
                - {$ref: "#/components/parameters/idPath"}
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required: [policy_id]
                            properties:
                                policy_id: {type: string}
            responses:
                "200":
                    description: Policy assigned
        delete:
            tags: [DoD]
            summary: Unassign DoD policy from endeavour
            parameters:
                - {$ref: "#/components/parameters/idPath"}
            responses:
                "200":
                    description: Policy unassigned
...
    /api/v1/organizations/{id}/export:
        get:
            tags: [Organizations]
            summary: Export organization data
            parameters:
                - {$ref: "#/components/parameters/idPath"}
            responses:
                "200":
                    description: Full organization export (JSON)
...